POSIX threads, mutexes and semaphores
In this  project,I am create a simulation of a cafe using POSIX threads, mutexes and semaphores. Your program should have the following functionalities and operations.
Your cafe should only sell one product, a normal coffee to make it simpler. The program scenario is, consumers come to the cafe and tries to find an empty register to buy a coffee.
Registers take the order, prepare it and give it to the customer and the customer leaves.
Your program should contain global static variables to manage your code.